Code Refactoring for better readability
