+++ screensaver_app/PasswordConfig.py
@@ -11,7 +11,7 @@ sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from central_logger import get_logger, log_startup, log_shutdown, log_exception
 logger = get_logger('PasswordConfig')
 
-from config_utils import find_user_config_path
+from utils.config_utils import find_user_config_path
 DEFAULT_PASSWORD = "1234"
 DEFAULT_USERNAME = "User" # Default for creating a new config if none exists
