+++ screensaver_app/widgets/media_widget.py
