+++ screensaver_app/widgets/stock_widget.py
