--- themes.py
@@ -1,25 +1,0 @@
-# Theme definitions
-
-# Dark theme (default)
-dark_theme = {
-    "bg": "#1e1e1e",
-    "fg": "#ffffff",
-    "button_bg": "#3d3d3d",
-    "button_fg": "#ffffff",
-    "listbox_bg": "#2d2d2d",
-    "listbox_fg": "#ffffff",
-    "highlight_bg": "#0078d7",
-    "highlight_fg": "#ffffff"
-}
-
-# Light theme
-light_theme = {
-    "bg": "#f0f0f0",
-    "fg": "#000000",
-    "button_bg": "#e1e1e1",
-    "button_fg": "#000000",
-    "listbox_bg": "#ffffff",
-    "listbox_fg": "#000000",
-    "highlight_bg": "#0078d7",
-    "highlight_fg": "#ffffff"
-}
